% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurelier-package.R
\docType{package}
\name{aurelier-package}
\alias{aurelier}
\alias{aurelier-package}
\title{aurelier: Create R Package In Autonomy}
\description{
Autonomy session with ThinkR.
}
\author{
\strong{Maintainer}: Aurelie Kamoun \email{aurelie.kamoun@finddx.org} (\href{https://orcid.org/0000-0001-8403-1944}{ORCID})

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
