% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_integrity.R
\name{check_data_integrity}
\alias{check_data_integrity}
\title{Title}
\usage{
check_data_integrity(dataset)
}
\arguments{
\item{dataset}{a dataset with column primary_fur_color that the function will check}
}
\value{
outputs message in the console to specify if the integrity check was successful
}
\description{
Title
}
